Revert confirm deploy project modal change
